+++ profiles/preview/preview/videoanalyzer/mgmt/videoanalyzer/videoanalyzerapi/models.go
@@ -9,11 +9,22 @@
 
 package videoanalyzerapi
 
-import original "github.com/Azure/azure-sdk-for-go/services/preview/videoanalyzer/mgmt/2021-05-01-preview/videoanalyzer/videoanalyzerapi"
+import original "github.com/Azure/azure-sdk-for-go/services/preview/videoanalyzer/mgmt/2021-11-01-preview/videoanalyzer/videoanalyzerapi"
 
 type AccessPoliciesClientAPI = original.AccessPoliciesClientAPI
 type EdgeModulesClientAPI = original.EdgeModulesClientAPI
+type LivePipelineOperationStatusesClientAPI = original.LivePipelineOperationStatusesClientAPI
+type LivePipelinesClientAPI = original.LivePipelinesClientAPI
 type LocationsClientAPI = original.LocationsClientAPI
+type OperationResultsClientAPI = original.OperationResultsClientAPI
+type OperationResultsGroupClientAPI = original.OperationResultsGroupClientAPI
+type OperationStatusesClientAPI = original.OperationStatusesClientAPI
+type OperationStatusesGroupClientAPI = original.OperationStatusesGroupClientAPI
 type OperationsClientAPI = original.OperationsClientAPI
+type PipelineJobOperationStatusesClientAPI = original.PipelineJobOperationStatusesClientAPI
+type PipelineJobsClientAPI = original.PipelineJobsClientAPI
+type PipelineTopologiesClientAPI = original.PipelineTopologiesClientAPI
+type PrivateEndpointConnectionsClientAPI = original.PrivateEndpointConnectionsClientAPI
+type PrivateLinkResourcesClientAPI = original.PrivateLinkResourcesClientAPI
 type VideoAnalyzersClientAPI = original.VideoAnalyzersClientAPI
 type VideosClientAPI = original.VideosClientAPI