+++ sdk/messaging/azservicebus/message.go
@@ -32,9 +32,6 @@ type (
 
 		rawAMQPMessage *amqp.Message
 
-		// the 'revision' of the link we received on (ie, if it was recovered it won't match)
-		linkRevision uint64
-
 		// deferred indicates we received it using ReceiveDeferredMessages. These messages
 		// will still go through the normal Receiver.Settle functions but internally will
 		// always be settled with the management link.